Ptl 1.5 support (#115)
Various changes to support Pytorch Lightning 1.5.

Major change is the removal of the `RayClusterEnvironment`. In PTL 1.5, the plugins leverage the cluster environment more heavily for specifying node rank, local rank, global rank, world size. With the RayPlugin, however, the ranks are not known until after the actors are launched and training begins. Thus, the cluster environment ended up adding more complexity than needed, and this PR removes the cluster environment and handles all of it internally in the plugin.

Other changes include supporting various name changes, refactoring, and signature changes in the DDPSpawnPlugin and DDPSpawnShardedPlugin.

Closes #108, #101

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>