Fix hanging `trainer.test()` (#142)
Closes #132

Fixes an issue where trainer.test() hangs when using multiple workers in the test DataLoader.

This issue is a bit weird as I was only able to reproduce the hanging with that exact setup. It does not occur with trainer.train() and multiple workers in the training DataLoader. It also does not occur when I set the num_workers for the test DataLoader to 0.

I'm not exactly sure what's going on, but the testing actually finishes and the program hangs at torch.distributed.destroy_process_group(). It may potentially be related to pytorch/pytorch#75097. The difference between trainer.train() and trainer.test() is that the former wraps the model in DDP while the latter doesn't (but still creates the process group).

In any case, the shudown_remote cleanup code is not actually necessary- the CUDA cache cleanup is already being called in the parent DDPSpawnPlugin on each worker, and it seems that torch.distributed.destroy_process_group() is not a Public API (and is not being called by PyTorch Lightning either).

The test added in this PR hangs prior to the changes to ray_ddp.py, but is passing after.